+++ auto_route_generator/CHANGELOG.md
@@ -1,11 +1,18 @@
 # ChangeLog
 
+## [0.4.4]
+
+- Change generating arguments holder for single parameter routes is now optional
+- Fix android soft back button always pops root navigator
+
 ## [0.4.3] Breaking Changes
+
 - Add ability to pass initial route arguments to ExtendedNavigator
 - Change single route parameters will have argument holder classes too as requested
 - Fix ExtendedNavigator.ofRouter<T>() returns null in inspector mode
 
 ## [0.4.0] Breaking Changes
+
 - Change using ExtendedNavigator instead of the native Navigator
 - Fix initial route does not go through guards
 - Change generated router class is no longer static